Added: New service for communication between components
