+++ src/app/components/player/player.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { BackendService } from 'src/app/services/backend.service';
+import { CommunicationService } from 'src/app/services/communication.service';
 
 @Component({
   selector: 'player',
@@ -9,12 +10,12 @@ import { BackendService } from 'src/app/services/backend.service';
 export class PlayerComponent implements OnInit {
 
   private audioPlayer: HTMLAudioElement;
-  constructor(private backendService: BackendService) { }
+  constructor(private backendService: BackendService, private communicationService: CommunicationService) { }
   songName: string;
   ngOnInit() {
     this.audioPlayer = document.getElementById('audioPlayer') as HTMLAudioElement;
     this.audioPlayer.volume = 0.3;
-    this.backendService.GetPlaySongObservable().subscribe(x => {
+    this.communicationService.GetPlaySongObservable().subscribe(x => {
       this.songName = x.name;
       this.audioPlayer.src = this.backendService.GetAudioStreamURL(x.songId);
       this.audioPlayer.play();
